+++ packaging/homebrew/StudyBar.rb
@@ -1,7 +1,7 @@
 cask "studybar" do
-  version "1.7.5"
+  version "1.7.57"
 
-  sha256 "80efd75685e06ce965ae409932dde673bdfd83a79612650309fde6569f8cf51e"
+  sha256 "87c01bbc3e542a4ce3c9b9d9ff69f6ec8ab323ff0aa8134784423a8da0f0aac4"
 
   url "https://github.com/IMisbahk/studybar/releases/download/v#{version}/StudyBar-#{version}.dmg"
   name "StudyBar"